pyproject.toml
[project]
name = "customer-support"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.2.0",
    "langchain>=1.2.15",
    "langchain-google-genai>=4.2.2",
    "langgraph>=1.1.8",
    "python-dotenv>=1.2.2",
]
